fix(uri): correct parsing for user:pass
correctly identify a user:pass@ connection string in a url, and do not
misidentify it as an alias. Also, adds a fix for ssh style string which
do not contain the `@` symbol, but only a domain.

To be clear, an alias cannot contain a `.` character, and for clarity
must at least specify a full domain, which will be checked in future
code.

Also includes the snapshots for the tests, which were accidentally
omitted from the previous commit.