feat(publish): collect paths passed on cli
This commit implements the user facing portion of the `eka publish`
subcommand. Users can pass paths relative to their in repo location,
or absolute paths relative to the repository root.

Warnings will be printed for paths that cannot be published because
they do not exist in the specified `--ref`.

Internally, this commit determines the object identity of an atom's
manifest and its directory of files within said history.