feat: fully functional publish subcommand
With this commit, the `eka publish` subcommand is fully functional and
creates a totally self-contained reference to an orphaned commit in the
repository, pointing to the specified atom.

All the critical pieces are included:
* The atom manifest is verified for validity
* The atom can be published with or without a companion directory
* The commit formulating the atom is reproducible
* The tree is trivially verifiable from the origin commit

Some minor implementation details may be changed. For example, currently
the author and commiter are left empty, but perhaps we could decide on
some constant or convention before stabalization.