feat: refactor published atom to two refs
In order for the later `eka resolve` command to function efficiently,
we need to be able to fetch manifest data as quickly as possible.

Before the version resolver has ran, all it has access to is the atom
path in the source repository as well as its requested version.

The resolver will be able to read all of the available atoms in a given
repository efficiently, and select one based on the request in the
manifest, but once it does it needs to be able to grab its manifest
as quickly as possible, to continue its work.

With one ref pointing at the atom's source, we cannot guarantee this,
since we would have to fetch the whole source of the atom to get just
the manifest and potential lock file.

So instead we split atom's up into two commits and two refs of the form:
refs/atoms/<path-in-repo>/<atom-name>-<version>/{manifest,source}