feat: add a static musl build & release profile
This commits adds a statically linked musl build via the
`x86_64-unknown-linux-musl` target. When building this target, the build
will automatically be fully statically linked.

In addition, the linker used under the hood is `mold` for all build
targets, because it works for both the dynamic and static builds
equally nicely, and it is quite fast.

In additional, a release profile including explicit optimization
settings is delcared in the Cargo.toml. Since tracing will preserve any
information we want in production, we opt to strip the symbols from the
binary to reduce its size.