refactor: fix URI parser for new atom id semantics
We will no longer be addressing atom's by their path in repo, as this
information is incidental at best. Instead we will address them in the
uri by their id specced in the manifest as:
```toml
[atom]
id = "my-atom"
```

This changes the delimiter between the url portion and the atom id
to `::`, since the remaining uri is no longer a path.