refactor(uri): major overhaul
Parsing logic is greatly simplified, and probably also more performant.

This is mostly because we first split the url and atom id portions of
the url based on the unique `::` delimiter, meaning that any conflicting
characters between the two are no longer an issue, since we can parse
them separately. This should also end up being more performant and
maintainable as well.

In addition, it is possible to now use the and resolve url aliases
wherever urls are expected on the cli, since the two paths have been
decoupled.