refactor: leverage parser improvements
The gix url crate has some peculiarities that make the previous method
of building urls into a string problematic. We can write more
maintainable code if we simply build the url with the builder function.

This requires writing some extra parsers to grab a possible port, as
well as unambiguously distinguishing the host, but the benefit is that
we don't need to maintain convoluted code paths to try and determine
the url type.