fix(id): calculate AtomId trivially
We have submitted an upstream patch to allow the iteration over git
histories to be a bit more versatile upstream:
https://github.com/Byron/gitoxide/pull/1610

In particular this change introduces the ability to traverse using a
oldest commit first strategy, reducing the time it takes to determine
the repositories root commit to a few milliseconds as opposed to the
several seconds in large histories.

This change is critical to maintain, which is the point of trying to
push the expanded API upstream, as it makes calculating and addressing
our atoms trivial and inexpensive.