feat(publish): introduce builder API
In order to ensure that is impossible for a user to publish when two
atoms in the same worktree share the same ID, we need to make the
constructor for the types that implement the `Publish` trait totally
private.

Instead we only expose them through a new `Builder` API that will run
a validation over the worktree to ensure that there are absolutely no
duplicates in the requested revspec, or fail.

Users may recover from this failure as desired, however, they will never
be able to construct a type with the `publish` method unless they fix
the duplicates, since the type can now only be constructed through this
builder, and the builder always runs this validation.