fix: don't use atom id in refspec
Using the full atom ID hash in the refspec obscures the name and makes
atoms impossible to search without some form of external index. This
isn't something we wish to encourage, as we want atom's to be self
contained and require no external dependencies other than the store in
which they are published.