feat!: change atom manifest format semantic
Instead of using a custom file extension for a toml file, which will
break developer tooling, and require manual intervent, instead we will
stick with the .toml suffix, but enforce a unique rule. Any toml file
ending in @.toml will be considered an atom.

This is particularly nice since it is short and concise, and also
reads naturally, e.g.: dev@.toml is the developement Atom TOML