fix: correct conditional compilation
gix is a somewhat heavy dependency, in addition, as the atom and eka api
evolve, it will be quite possible that not all users will want or need
all possible features. It was always intended to keep the major feature
groups guarded behind features (such as store implementations), but we
sort of fell behind keeping track of that for a bit.

This commit corrects that and ensures that compilation works both with
or without features enabled.